Split fullscreen TUI rendering module
